Add files via upload
Velocity calculating script that uses two sets of points of data of the same points but at different times.

Points msg contains the structure for transporting the sets of points. Time1, Xs1, Ys1 for the newer data and Time2, Xs2, and Ys2 for the older data.

use "from terrapin_points_msg.msg import points" to use.